pyproject.toml
[tool.poetry]
name = "epipipeline"
version = "2.0.13"
description = ""
authors = ["Sai Sneha <sneha@artpark.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.0"
boto3 = "^1.34.40"
requests = "^2.31.0"
fuzzywuzzy = "^0.18.0"
openpyxl = "^3.1.2"
python-levenshtein = "^0.25.0"
pyyaml = "^6.0.1"
ipython = "^8.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
